Various Visual Studio work space updates: - remove vc8 - rename vc10 -> vc2010 - vc2010: fix include paths and platform settings - fix a warning for 64-bit builds on Windows
Note: Batch Build does not work on VS2010, except for the "current" platform

+++ branches/HM-16.6-JEM-2.0-dev/source/Lib/TLibEncoder/TEncSearch.cpp
@@ -732,7 +732,7 @@ Void TEncSearch::init(TEncCfg*      pcEncCfg,
       {
           for (Int hidx=0; hidx<uiNumLayersToAllocate; hidx++)
           {
-              m_resiSingleBuffer[widx][hidx][n] = new Pel [1<<(widx+hidx+(MIN_CU_LOG2<<1))];
+              m_resiSingleBuffer[widx][hidx][n] = new Pel [(size_t)1<<(widx+hidx+(MIN_CU_LOG2<<1))];
           }
       }
 #else
